# src/build/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"ApiResponse_any_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"data": {},
					"message": {
						"type": "string"
					},
					"error": {
						"type": "string"
					},
					"details": {},
					"pagination": {
						"properties": {
							"totalPages": {
								"type": "number",
								"format": "double"
							},
							"total": {
								"type": "number",
								"format": "double"
							},
							"limit": {
								"type": "number",
								"format": "double"
							},
							"page": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"totalPages",
							"total",
							"limit",
							"page"
						],
						"type": "object"
					}
				},
				"required": [
					"success"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"TitleResponse": {
				"properties": {
					"title_id": {
						"type": "string"
					},
					"course_id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					}
				},
				"required": [
					"title_id",
					"course_id",
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_TitleResponse-Array_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/TitleResponse"
						},
						"type": "array"
					},
					"message": {
						"type": "string"
					},
					"error": {
						"type": "string"
					},
					"details": {},
					"pagination": {
						"properties": {
							"totalPages": {
								"type": "number",
								"format": "double"
							},
							"total": {
								"type": "number",
								"format": "double"
							},
							"limit": {
								"type": "number",
								"format": "double"
							},
							"page": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"totalPages",
							"total",
							"limit",
							"page"
						],
						"type": "object"
					}
				},
				"required": [
					"success"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_TitleResponse_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"data": {
						"$ref": "#/components/schemas/TitleResponse"
					},
					"message": {
						"type": "string"
					},
					"error": {
						"type": "string"
					},
					"details": {},
					"pagination": {
						"properties": {
							"totalPages": {
								"type": "number",
								"format": "double"
							},
							"total": {
								"type": "number",
								"format": "double"
							},
							"limit": {
								"type": "number",
								"format": "double"
							},
							"page": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"totalPages",
							"total",
							"limit",
							"page"
						],
						"type": "object"
					}
				},
				"required": [
					"success"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateTitleRequest": {
				"properties": {
					"course_id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					}
				},
				"required": [
					"course_id",
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateTitleRequest": {
				"properties": {
					"title_id": {
						"type": "string"
					},
					"course_id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					}
				},
				"required": [
					"title_id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"TestResultResponse": {
				"properties": {
					"result_id": {
						"type": "string"
					},
					"user_id": {
						"type": "string"
					},
					"test_id": {
						"type": "string"
					},
					"start_date": {
						"type": "string",
						"format": "date-time"
					},
					"end_date": {
						"type": "string",
						"format": "date-time"
					},
					"score": {
						"type": "number",
						"format": "double"
					},
					"ai_feedback": {
						"type": "string"
					}
				},
				"required": [
					"result_id",
					"user_id",
					"test_id",
					"start_date"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_TestResultResponse-Array_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/TestResultResponse"
						},
						"type": "array"
					},
					"message": {
						"type": "string"
					},
					"error": {
						"type": "string"
					},
					"details": {},
					"pagination": {
						"properties": {
							"totalPages": {
								"type": "number",
								"format": "double"
							},
							"total": {
								"type": "number",
								"format": "double"
							},
							"limit": {
								"type": "number",
								"format": "double"
							},
							"page": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"totalPages",
							"total",
							"limit",
							"page"
						],
						"type": "object"
					}
				},
				"required": [
					"success"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_TestResultResponse_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"data": {
						"$ref": "#/components/schemas/TestResultResponse"
					},
					"message": {
						"type": "string"
					},
					"error": {
						"type": "string"
					},
					"details": {},
					"pagination": {
						"properties": {
							"totalPages": {
								"type": "number",
								"format": "double"
							},
							"total": {
								"type": "number",
								"format": "double"
							},
							"limit": {
								"type": "number",
								"format": "double"
							},
							"page": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"totalPages",
							"total",
							"limit",
							"page"
						],
						"type": "object"
					}
				},
				"required": [
					"success"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateTestResultRequest": {
				"properties": {
					"user_id": {
						"type": "string"
					},
					"test_id": {
						"type": "string"
					},
					"start_date": {
						"type": "string",
						"format": "date-time"
					},
					"end_date": {
						"type": "string",
						"format": "date-time"
					},
					"score": {
						"type": "number",
						"format": "double"
					},
					"ai_feedback": {
						"type": "string"
					}
				},
				"required": [
					"user_id",
					"test_id",
					"start_date"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateTestResultRequest": {
				"properties": {
					"result_id": {
						"type": "string"
					},
					"user_id": {
						"type": "string"
					},
					"test_id": {
						"type": "string"
					},
					"start_date": {
						"type": "string",
						"format": "date-time"
					},
					"end_date": {
						"type": "string",
						"format": "date-time"
					},
					"score": {
						"type": "number",
						"format": "double"
					},
					"ai_feedback": {
						"type": "string"
					}
				},
				"required": [
					"result_id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"TestResponse": {
				"properties": {
					"test_id": {
						"type": "string"
					},
					"course_id": {
						"type": "string"
					},
					"type_status": {
						"type": "string"
					},
					"title": {
						"type": "string"
					}
				},
				"required": [
					"test_id",
					"type_status",
					"title"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_TestResponse-Array_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/TestResponse"
						},
						"type": "array"
					},
					"message": {
						"type": "string"
					},
					"error": {
						"type": "string"
					},
					"details": {},
					"pagination": {
						"properties": {
							"totalPages": {
								"type": "number",
								"format": "double"
							},
							"total": {
								"type": "number",
								"format": "double"
							},
							"limit": {
								"type": "number",
								"format": "double"
							},
							"page": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"totalPages",
							"total",
							"limit",
							"page"
						],
						"type": "object"
					}
				},
				"required": [
					"success"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_TestResponse_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"data": {
						"$ref": "#/components/schemas/TestResponse"
					},
					"message": {
						"type": "string"
					},
					"error": {
						"type": "string"
					},
					"details": {},
					"pagination": {
						"properties": {
							"totalPages": {
								"type": "number",
								"format": "double"
							},
							"total": {
								"type": "number",
								"format": "double"
							},
							"limit": {
								"type": "number",
								"format": "double"
							},
							"page": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"totalPages",
							"total",
							"limit",
							"page"
						],
						"type": "object"
					}
				},
				"required": [
					"success"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateTestRequest": {
				"properties": {
					"course_id": {
						"type": "string"
					},
					"type_status": {
						"type": "string"
					},
					"title": {
						"type": "string"
					}
				},
				"required": [
					"type_status",
					"title"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateTestRequest": {
				"properties": {
					"test_id": {
						"type": "string"
					},
					"course_id": {
						"type": "string"
					},
					"type_status": {
						"type": "string"
					},
					"title": {
						"type": "string"
					}
				},
				"required": [
					"test_id"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "express-pg-orm-backend-ts",
		"version": "1.0.0",
		"description": "TypeScript Express.js backend with PostgreSQL and Sequelize ORM",
		"contact": {}
	},
	"paths": {
...
		"/titles/list": {
			"post": {
				"operationId": "ListTitles",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_TitleResponse-Array_"
								}
							}
						}
					}
				},
				"tags": [
					"称号表"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"limit": {
										"type": "number",
										"format": "double"
									},
									"page": {
										"type": "number",
										"format": "double"
									}
								},
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/titles/getById": {
			"post": {
				"operationId": "GetTitleById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_TitleResponse_"
								}
							}
						}
					}
				},
				"tags": [
					"称号表"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"title_id": {
										"type": "string"
									}
								},
								"required": [
									"title_id"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/titles/add": {
			"post": {
				"operationId": "AddTitle",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_any_"
								}
							}
						}
					}
				},
				"tags": [
					"称号表"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateTitleRequest"
							}
						}
					}
				}
			}
		},
		"/titles/update": {
			"post": {
				"operationId": "UpdateTitle",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_any_"
								}
							}
						}
					}
				},
				"tags": [
					"称号表"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateTitleRequest"
							}
						}
					}
				}
			}
		},
		"/titles/delete": {
			"post": {
				"operationId": "DeleteTitle",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_any_"
								}
							}
						}
					}
				},
				"tags": [
					"称号表"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"title_id": {
										"type": "string"
									}
								},
								"required": [
									"title_id"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/test-results/list": {
			"post": {
				"operationId": "ListTestResults",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_TestResultResponse-Array_"
								}
							}
						}
					}
				},
				"tags": [
					"测试结果表"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"limit": {
										"type": "number",
										"format": "double"
									},
									"page": {
										"type": "number",
										"format": "double"
									}
								},
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/test-results/getById": {
			"post": {
				"operationId": "GetTestResultById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_TestResultResponse_"
								}
							}
						}
					}
				},
				"tags": [
					"测试结果表"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"result_id": {
										"type": "string"
									}
								},
								"required": [
									"result_id"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/test-results/add": {
			"post": {
				"operationId": "AddTestResult",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_any_"
								}
							}
						}
					}
				},
				"tags": [
					"测试结果表"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateTestResultRequest"
							}
						}
					}
				}
			}
		},
		"/test-results/update": {
			"post": {
				"operationId": "UpdateTestResult",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_any_"
								}
							}
						}
					}
				},
				"tags": [
					"测试结果表"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateTestResultRequest"
							}
						}
					}
				}
			}
		},
		"/test-results/delete": {
			"post": {
				"operationId": "DeleteTestResult",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_any_"
								}
							}
						}
					}
				},
				"tags": [
					"测试结果表"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"result_id": {
										"type": "string"
									}
								},
								"required": [
									"result_id"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/tests/list": {
			"post": {
				"operationId": "ListTests",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_TestResponse-Array_"
								}
							}
						}
					}
				},
				"tags": [
					"测试表"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"limit": {
										"type": "number",
										"format": "double"
									},
									"page": {
										"type": "number",
										"format": "double"
									}
								},
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/tests/getById": {
			"post": {
				"operationId": "GetTestById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_TestResponse_"
								}
							}
						}
					}
				},
				"tags": [
					"测试表"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"test_id": {
										"type": "string"
									}
								},
								"required": [
									"test_id"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/tests/add": {
			"post": {
				"operationId": "AddTest",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_any_"
								}
							}
						}
					}
				},
				"tags": [
					"测试表"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateTestRequest"
							}
						}
					}
				}
			}
		},
		"/tests/update": {
			"post": {
				"operationId": "UpdateTest",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_any_"
								}
							}
						}
					}
				},
				"tags": [
					"测试表"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateTestRequest"
							}
						}
					}
				}
			}
		},
		"/tests/delete": {
			"post": {
				"operationId": "DeleteTest",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_any_"
								}
							}
						}
					}
				},
				"tags": [
					"测试表"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"test_id": {
										"type": "string"
									}
								},
								"required": [
									"test_id"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}